+++ video2gif.sh
@@ -1,9 +1,7 @@
 #!/bin/bash
 
 PROG=video2gif
-VERSION=1.0.0
-GIF_HEIGHT=480
-GIF_FPS=12
+VERSION=1.1.0
 
 usage() {
 	echo >&2 "usage: $PROG [-h] VIDEO_FILE"
@@ -13,15 +11,19 @@ print_help() {
 	usage
 	cat >&2 <<-END
 
-	Convert a video file to a GIF. Resulting GIF will have a height
-	of 480 pixels, framerate of 12 FPS, and the same base name as
-	VIDEO_FILE, but with ".gif" extension.
+	Convert a video file to a GIF. Resulting GIF will have framerate of 12
+	FPS and the same base name as VIDEO_FILE, but with ".gif" extension.
 
 	positional arguments:
 	  VIDEO_FILE    a path to the video file you want to convert
 
 	optional arguments:
+	  -f, --framerate FPS
+	                 set the animation framerate of the GIF (default: 12)
 	  -h, --help     display this help and exit
+	  -y, --height HEIGHT
+	                 set the height of the GIF, keep aspect ratio
+			 (default: same as video)
 	  -V, --version  output version information and exit
 	END
 }
@@ -38,11 +40,37 @@ usage_error() {
 	error "$message"
 }
 
+scale_filter=
+gif_height=
+gif_fps=12
+
 while :; do
 	case "$1" in
+		-f|--framerate)
+			[[ $# -lt 2 ]] && usage_error "argument $1: expected one argument"
+			gif_fps=$2
+			shift
+			;;
+		--framerate=*?)
+			gif_fps="${1#*=}"
+			;;
+		--framerate=)
+			usage_error '"--framerate" requires a non-empty option argument'
+			;;
 		-h|--help)
 			print_help
 			exit
+			;;
+		-y|--height)
+			[[ $# -lt 2 ]] && usage_error "argument $1: expected one argument"
+			gif_height=$2
+			shift
+			;;
+		--height=*?)
+			gif_height="${1#*=}"
+			;;
+		--height=)
+			usage_error '"--height" requires a non-empty option argument'
 			;;
 		-V|--version)
 			echo $VERSION && exit
@@ -56,14 +84,16 @@ while :; do
 	shift
 done
 
+[[ "$gif_height" ]] && scale_filter=",scale=-1:$gif_height"
+
 video_file="$1"
-if [[ -z "$1" ]]; then
+if [[ -z "$video_file" ]]; then
 	usage_error 'the following arguments are required: VIDEO_FILE'
 fi
 
 basename="${video_file%.*}"
 gif_file="${basename}.gif"
-gif_filter="[0:v] fps=${gif_fps},scale=${gif_height}:-1,split [a][b];"
+gif_filter="[0:v] fps=${gif_fps}${scale_filter},split [a][b];"
 gif_filter+='[a] palettegen [p];[b][p] paletteuse'
 
 ffmpeg -i "$video_file" \