+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
+## video2gif 1.1.0 - 2023-04-14
+
+### Added
+- `--framerate` option
+- `--height` option
+
 ## overlaysubs 1.1.0 - 2023-02-23
+- Height of the GIF was not being set to 480 pixels, as it was
+  promised in the help text, instead it was the WIDTH that was
+  always set to 480 pixels.
 
 ### Added
 - `--start-at` and `--end-at` options: include only a part of the input video in